dom/tests/mochitest/cowl/postmessage/test_postmessage_tainted.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>
  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var completedTests = 0;
  var passedTests = 0;

  function ok_wrapper(result, desc) {
    ok(result, desc);

    completedTests++;

    if (result) {
      passedTests++;
    }

    if (completedTests == 3) {
      is(passedTests, completedTests, "There are " + completedTests + " privilege tests that should pass");
      SimpleTest.finish();
    }
  }

  window.addEventListener('message', onReceive, false);
  function onReceive() {
    ok_wrapper(true, "Received a msg from context with priv");
  }

  </script>
  <iframe src="http://example.com/tests/dom/tests/mochitest/cowl/file_postmessage_tainted.html"></iframe>
  <iframe src="http://example.com/tests/dom/tests/mochitest/cowl/file_postmessage_from_tainted_w_priv.html"></iframe>
  <iframe src="http://example.com/tests/dom/tests/mochitest/cowl/file_postmessage_from_tainted_to_priv.html"></iframe>
</pre>
</body>
</html>
